+++ CHANGES.rst
@@ -105,6 +105,9 @@ Changelog
 - Make the `review_state_title` column (that translates the review_state title
   instead id) also available for dashboards displaying meetings.
   [gbastien]
+- Make sure meeting fieldsets order is correct when a custom field has been
+  added to an existing fieldset.
+  [gbastien]
 
 4.2rc34 (2022-09-29)
 --------------------